String driver needs to access stream
Overthought this - we need to use the stream since it is possible for
the consumer to directly access the stream (i.e. add filters to the
stream). So can't do shortcuts on PHP variables directly.

This is more proof-of-concept than anything though.  This was originally
intended to help performance of the Temp driver, but optimizations to
the base Horde_Stream code has pretty much solved the performance issues
we were previously seeing.